fixing some bugs

+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: mazhari <mazhari@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/09 14:26:34 by mazhari           #+#    #+#             */
-/*   Updated: 2022/03/28 20:54:57 by mazhari          ###   ########.fr       */
+/*   Updated: 2022/04/01 16:06:59 by mazhari          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ typedef struct s_philo
 	pthread_mutex_t	eat;
 	int				nbr_eat;
 	int				nbr;
-	long			death_time;
+	long long		death_time;
 	struct s_data	*data;
 }				t_philo;
 
@@ -39,7 +39,7 @@ typedef struct s_data
 	int				time_to_eat;
 	int				time_to_sleep;
 	int				nbr_must_eat;
-	long			start_time;
+	long long		start_time;
 	int				death;
 	int				finish_eat;
 	pthread_mutex_t	print;
@@ -49,7 +49,7 @@ int		ft_atoi(const char *str);
 int		data_init(t_data *data, char **av);
 void	*philosopher(void *p);
 int		exit_program(t_data *data);
-long	get_time(void);
+long long	get_time(void);
 void	print(t_philo philo, char *str);
 
 #endif
